+++ Src/Opinion/Rating/Rating_Handler.cpp
@@ -11,8 +11,8 @@ void Rating_Handler::delete_previous_rating(std::string writer_) {
 		ratings.end());
 }
 
-Rating* Rating_Handler::add_rating(std::string writer_, float location_, float cleanness_, float staff_, float facilities_,
-                                float value_for_money_, float overall_) {
+Rating* Rating_Handler::add_rating(std::string writer_, double location_, double cleanness_, double staff_, double facilities_,
+                                double value_for_money_, double overall_) {
 
 	Rating* rating = new Rating(writer_,location_,cleanness_,staff_,facilities_,value_for_money_,overall_);
 	delete_previous_rating(writer_);
@@ -27,7 +27,7 @@ void Rating_Handler::add_rating(std::string hotel_id, Rating* rating) {
 	_ratings_.insert(std::pair<std::string,Rating*>(hotel_id,rating));
 }
 
-float Rating_Handler::do_rated(std::string hotel_id) {
+double Rating_Handler::do_rated(std::string hotel_id) {
 
 	auto item = _ratings_.find(hotel_id);
 	if(item != _ratings_.end()) return item->second->overall;
